Extract HCC data into CSV.
Shuffle around some code to make it reusable.
